tutorials/02-weather-vibes-agent/weather_vibes/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
jinja2==3.1.2
pydantic>=2.6.1,<3.0.0
rich>=13.0.0

# OpenAI
openai>=1.0.0

# API integrations
requests==2.31.0
google-api-python-client==2.111.0

# Evaluation and monitoring
#  uncomment this requirement as part of Step 7 to use Galileo.
# galileo

# Simple Agent Framework
git+https://github.com/rungalileo/simple-agent-framework.git@main 